Split app-server tests into modules.
