+++ components/app_server/src/lib.rs
@@ -5,9 +5,11 @@
 #![feature(generators)]
 #![feature(never_type)]
 
-
 #[macro_use]
 extern crate log;
 
 mod server;
 mod config;
+
+#[cfg(test)]
+mod tests;
