+++ components/app_server/src/server.rs
@@ -20,7 +20,7 @@ use proto::index_client::messages::{IndexClientToAppServer, AppServerToIndexClie
 
 use crate::config::AppPermissions;
 
-type IncomingAppConnection<B,ISA> = (AppPermissions, ConnPair<AppServerToApp<B,ISA>, AppToAppServer<B,ISA>>);
+pub type IncomingAppConnection<B,ISA> = (AppPermissions, ConnPair<AppServerToApp<B,ISA>, AppToAppServer<B,ISA>>);
 
 
 #[derive(Debug)]
@@ -455,588 +455,4 @@ where
         }
     }
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    use futures::executor::ThreadPool;
-    use futures::{FutureExt, TryFutureExt};
-
-    use im::hashmap::HashMap as ImHashMap;
-
-    use crypto::identity::{PublicKey, PUBLIC_KEY_LEN};
-    use crypto::uid::UID_LEN;
-    use proto::funder::report::{FunderReport, FunderReportMutation};
-    use proto::index_client::messages::{IndexClientReport, IndexClientReportMutation, 
-        ClientResponseRoutes, ResponseRoutesResult};
-    use proto::index_server::messages::RequestRoutes;
-    use proto::funder::messages::{UserRequestSendFunds, FriendsRoute, 
-        InvoiceId, INVOICE_ID_LEN, ResponseReceived, ResponseSendFundsResult};
-
-    /// A test util function.
-    /// Spawns an app server loop and returns all relevant channels
-    /// used for control or communication.
-    fn spawn_dummy_app_server<S>(mut spawner: S) -> 
-        (mpsc::Sender<FunderOutgoingControl<Vec<u32>>>,
-         mpsc::Receiver<FunderIncomingControl<Vec<u32>>>,
-         mpsc::Sender<IndexClientToAppServer<u64>>,
-         mpsc::Receiver<AppServerToIndexClient<u64>>,
-         mpsc::Sender<IncomingAppConnection<u32,u64>>,
-         NodeReport<u32,u64>)
-
-    where
-        S: Spawn + Clone + Send + 'static,
-    {
-        let (mut funder_sender, from_funder) = mpsc::channel(0);
-        let (to_funder, mut funder_receiver) = mpsc::channel(0);
-
-        let (index_client_sender, from_index_client) = mpsc::channel(0);
-        let (to_index_client, index_client_receiver) = mpsc::channel(0);
-
-        let (mut connections_sender, incoming_connections) = mpsc::channel(0);
-
-        // Create a dummy initial_node_report:
-        let funder_report = FunderReport {
-            local_public_key: PublicKey::from(&[0xaa; PUBLIC_KEY_LEN]),
-            address: vec![0u32, 1u32],
-            friends: ImHashMap::new(),
-            num_ready_receipts: 0,
-        };
-
-        let index_client_report = IndexClientReport {
-            index_servers: vec![100u64, 101u64],
-            opt_connected_server: Some(101u64),
-        };
-
-        let initial_node_report = NodeReport {
-            funder_report,
-            index_client_report,
-        };
-
-        let fut_loop = app_server_loop(from_funder,
-                        to_funder,
-                        from_index_client,
-                        to_index_client,
-                        incoming_connections,
-                        initial_node_report.clone(),
-                        spawner.clone())
-            .map_err(|e| error!("app_server_loop() error: {:?}", e))
-            .map(|_| ());
-
-        spawner.spawn(fut_loop).unwrap();
-
-        (funder_sender, funder_receiver,
-         index_client_sender, index_client_receiver,
-         connections_sender, initial_node_report)
-
-
-    }
-
-    async fn task_app_server_loop_funder_command<S>(mut spawner: S) 
-    where
-        S: Spawn + Clone + Send + 'static,
-    {
-
-        let (mut funder_sender, mut funder_receiver,
-             index_client_sender, index_client_receiver,
-             mut connections_sender, initial_node_report) = spawn_dummy_app_server(spawner.clone());
-
-        let (mut app_sender, app_server_receiver) = mpsc::channel(0);
-        let (app_server_sender, mut app_receiver) = mpsc::channel(0);
-        let app_server_conn_pair = (app_server_sender, app_server_receiver);
-
-        let app_permissions = AppPermissions {
-            reports: true,
-            routes: true,
-            send_funds: true,
-            config: true,
-        };
-
-        await!(connections_sender.send((app_permissions, app_server_conn_pair))).unwrap();
-
-        // The app should receive the current node report as the first message:
-        let to_app_message = await!(app_receiver.next()).unwrap();
-        match to_app_message {
-            AppServerToApp::Report(report) => assert_eq!(report, initial_node_report),
-            _ => unreachable!(),
-        };
-
-        // Send a command through the app:
-        let new_address = vec![0u32, 1u32, 2u32];
-        await!(app_sender.send(AppToAppServer::SetRelays(new_address.clone()))).unwrap();
-
-        // SetRelays command should be forwarded to the Funder:
-        let to_funder_message = await!(funder_receiver.next()).unwrap();
-        match to_funder_message {
-            FunderIncomingControl::SetAddress(address) => assert_eq!(address, new_address),
-            _ => unreachable!(),
-        };
-
-        let funder_report_mutation = FunderReportMutation::SetAddress(new_address.clone());
-        let funder_report_mutations = vec![funder_report_mutation.clone()];
-        await!(funder_sender.send(FunderOutgoingControl::ReportMutations(funder_report_mutations))).unwrap();
-
-        let to_app_message = await!(app_receiver.next()).unwrap();
-        match to_app_message {
-            AppServerToApp::ReportMutations(report_mutations) => {
-                assert_eq!(report_mutations.len(), 1);
-                let report_mutation = &report_mutations[0];
-                match report_mutation {
-                    NodeReportMutation::Funder(received_funder_report_mutation) => {
-                        assert_eq!(received_funder_report_mutation, 
-                                   &funder_report_mutation);
-                    },
-                    _ => unreachable!(),
-                }
-            },
-            _ => unreachable!(),
-        }
-    }
-
-    #[test]
-    fn test_app_server_loop_funder_command() {
-        let mut thread_pool = ThreadPool::new().unwrap();
-        thread_pool.run(task_app_server_loop_funder_command(thread_pool.clone()));
-    }
-
-    // -------------------------------------------------
-    // -------------------------------------------------
-
-    async fn task_app_server_loop_index_client_command<S>(mut spawner: S) 
-    where
-        S: Spawn + Clone + Send + 'static,
-    {
-
-        let (mut funder_sender, mut funder_receiver,
-             mut index_client_sender, mut index_client_receiver,
-             mut connections_sender, initial_node_report) = spawn_dummy_app_server(spawner.clone());
-
-        let (mut app_sender, app_server_receiver) = mpsc::channel(0);
-        let (app_server_sender, mut app_receiver) = mpsc::channel(0);
-        let app_server_conn_pair = (app_server_sender, app_server_receiver);
-
-        let app_permissions = AppPermissions {
-            reports: true,
-            routes: true,
-            send_funds: true,
-            config: true,
-        };
-
-        await!(connections_sender.send((app_permissions, app_server_conn_pair))).unwrap();
-
-        // The app should receive the current node report as the first message:
-        let to_app_message = await!(app_receiver.next()).unwrap();
-        match to_app_message {
-            AppServerToApp::Report(report) => assert_eq!(report, initial_node_report),
-            _ => unreachable!(),
-        };
-
-        // Send a command through the app:
-        let new_index_server_address = 300u64;
-        await!(app_sender.send(AppToAppServer::AddIndexServer(new_index_server_address))).unwrap();
-
-        // AddIndexServer command should be forwarded to IndexClient:
-        let to_index_client_message = await!(index_client_receiver.next()).unwrap();
-        match to_index_client_message {
-            AppServerToIndexClient::AddIndexServer(address) => assert_eq!(address, new_index_server_address),
-            _ => unreachable!(),
-        };
-
-        let index_client_report_mutation = IndexClientReportMutation::AddIndexServer(new_index_server_address.clone());
-        let index_client_report_mutations = vec![index_client_report_mutation.clone()];
-        await!(index_client_sender.send(IndexClientToAppServer::ReportMutations(index_client_report_mutations))).unwrap();
-
-        let to_app_message = await!(app_receiver.next()).unwrap();
-        match to_app_message {
-            AppServerToApp::ReportMutations(report_mutations) => {
-                assert_eq!(report_mutations.len(), 1);
-                let report_mutation = &report_mutations[0];
-                match report_mutation {
-                    NodeReportMutation::IndexClient(received_index_client_report_mutation) => {
-                        assert_eq!(received_index_client_report_mutation, 
-                                   &index_client_report_mutation);
-                    },
-                    _ => unreachable!(),
-                }
-            },
-            _ => unreachable!(),
-        }
-    }
-
-    #[test]
-    fn test_app_server_loop_index_client_command() {
-        let mut thread_pool = ThreadPool::new().unwrap();
-        thread_pool.run(task_app_server_loop_index_client_command(thread_pool.clone()));
-    }
-
-
-    // -------------------------------------------------
-    // -------------------------------------------------
-
-    async fn task_app_server_loop_all_apps_closed<S>(mut spawner: S) 
-    where
-        S: Spawn + Clone + Send + 'static,
-    {
-
-        let (mut funder_sender, mut funder_receiver,
-             mut index_client_sender, mut index_client_receiver,
-             mut connections_sender, initial_node_report) = spawn_dummy_app_server(spawner.clone());
-
-        let (mut app_sender, app_server_receiver) = mpsc::channel(0);
-        let (app_server_sender, mut app_receiver) = mpsc::channel(0);
-        let app_server_conn_pair = (app_server_sender, app_server_receiver);
-
-        let app_permissions = AppPermissions {
-            reports: true,
-            routes: true,
-            send_funds: true,
-            config: true,
-        };
-
-        await!(connections_sender.send((app_permissions, app_server_conn_pair))).unwrap();
-
-        // The app should receive the current node report as the first message:
-        let to_app_message = await!(app_receiver.next()).unwrap();
-        match to_app_message {
-            AppServerToApp::Report(report) => assert_eq!(report, initial_node_report),
-            _ => unreachable!(),
-        };
-
-        drop(connections_sender);
-
-        // Communication with the last connected app should still work:
-        let new_index_server_address = 300u64;
-        let index_client_report_mutation = IndexClientReportMutation::AddIndexServer(new_index_server_address.clone());
-        let index_client_report_mutations = vec![index_client_report_mutation.clone()];
-        await!(index_client_sender.send(IndexClientToAppServer::ReportMutations(index_client_report_mutations))).unwrap();
-
-        let to_app_message = await!(app_receiver.next()).unwrap(); 
-
-        // Last app disconnects:
-        drop(app_sender);
-
-        // At this point the app_server loop should close.
-
-        assert!(await!(funder_receiver.next()).is_none());
-        assert!(await!(index_client_receiver.next()).is_none());
-    }
-
-    #[test]
-    fn test_app_server_loop_index_all_apps_closed() {
-        let mut thread_pool = ThreadPool::new().unwrap();
-        thread_pool.run(task_app_server_loop_all_apps_closed(thread_pool.clone()));
-    }
-
-
-    // -------------------------------------------------
-    // -------------------------------------------------
-
-    async fn task_app_server_loop_two_apps<S>(mut spawner: S) 
-    where
-        S: Spawn + Clone + Send + 'static,
-    {
-
-        let (mut funder_sender, mut funder_receiver,
-             mut index_client_sender, mut index_client_receiver,
-             mut connections_sender, initial_node_report) = spawn_dummy_app_server(spawner.clone());
-
-        let (mut app_sender0, app_server_receiver) = mpsc::channel(0);
-        let (app_server_sender, mut app_receiver0) = mpsc::channel(0);
-        let app_server_conn_pair = (app_server_sender, app_server_receiver);
-        let app_permissions = AppPermissions {
-            reports: true,
-            routes: true,
-            send_funds: true,
-            config: true,
-        };
-        await!(connections_sender.send((app_permissions, app_server_conn_pair))).unwrap();
-
-        let (mut app_sender1, app_server_receiver) = mpsc::channel(0);
-        let (app_server_sender, mut app_receiver1) = mpsc::channel(0);
-        let app_server_conn_pair = (app_server_sender, app_server_receiver);
-        let app_permissions = AppPermissions {
-            reports: true,
-            routes: true,
-            send_funds: true,
-            config: true,
-        };
-        await!(connections_sender.send((app_permissions, app_server_conn_pair))).unwrap();
-
-        // The apps should receive the current node report as the first message:
-        // Send a report 
-        let to_app_message = await!(app_receiver0.next()).unwrap();
-        match to_app_message {
-            AppServerToApp::Report(report) => assert_eq!(report, initial_node_report),
-            _ => unreachable!(),
-        };
-        let to_app_message = await!(app_receiver1.next()).unwrap();
-        match to_app_message {
-            AppServerToApp::Report(report) => assert_eq!(report, initial_node_report),
-            _ => unreachable!(),
-        };
-
-        // Send a dummy report message from IndexClient:
-        let new_index_server_address = 300u64;
-        let index_client_report_mutation = IndexClientReportMutation::AddIndexServer(new_index_server_address.clone());
-        let index_client_report_mutations = vec![index_client_report_mutation.clone()];
-        await!(index_client_sender.send(IndexClientToAppServer::ReportMutations(index_client_report_mutations))).unwrap();
-
-        // Both apps should get the report:
-        let to_app_message = await!(app_receiver0.next()).unwrap(); 
-        match to_app_message {
-            AppServerToApp::ReportMutations(report_mutations) => {
-                assert_eq!(report_mutations.len(), 1);
-                let report_mutation = &report_mutations[0];
-                match report_mutation {
-                    NodeReportMutation::IndexClient(received_index_client_report_mutation) => {
-                        assert_eq!(received_index_client_report_mutation, 
-                                   &index_client_report_mutation);
-                    },
-                    _ => unreachable!(),
-                }
-            },
-            _ => unreachable!(),
-        }
-        let to_app_message = await!(app_receiver1.next()).unwrap(); 
-        match to_app_message {
-            AppServerToApp::ReportMutations(report_mutations) => {
-                assert_eq!(report_mutations.len(), 1);
-                let report_mutation = &report_mutations[0];
-                match report_mutation {
-                    NodeReportMutation::IndexClient(received_index_client_report_mutation) => {
-                        assert_eq!(received_index_client_report_mutation, 
-                                   &index_client_report_mutation);
-                    },
-                    _ => unreachable!(),
-                }
-            },
-            _ => unreachable!(),
-        }
-    }
-
-    #[test]
-    fn test_app_server_loop_index_two_apps() {
-        let mut thread_pool = ThreadPool::new().unwrap();
-        thread_pool.run(task_app_server_loop_two_apps(thread_pool.clone()));
-    }
-
-    // ----------------------------------------------
-    // ----------------------------------------------
-
-
-    async fn task_app_server_loop_request_routes<S>(mut spawner: S) 
-    where
-        S: Spawn + Clone + Send + 'static,
-    {
-
-        let (mut funder_sender, mut funder_receiver,
-             mut index_client_sender, mut index_client_receiver,
-             mut connections_sender, initial_node_report) = spawn_dummy_app_server(spawner.clone());
-
-        // Connect two apps:
-        let (mut app_sender0, app_server_receiver) = mpsc::channel(0);
-        let (app_server_sender, mut app_receiver0) = mpsc::channel(0);
-        let app_server_conn_pair = (app_server_sender, app_server_receiver);
-        let app_permissions = AppPermissions {
-            reports: true,
-            routes: true,
-            send_funds: true,
-            config: true,
-        };
-        await!(connections_sender.send((app_permissions, app_server_conn_pair))).unwrap();
-
-        let (mut app_sender1, app_server_receiver) = mpsc::channel(0);
-        let (app_server_sender, mut app_receiver1) = mpsc::channel(0);
-        let app_server_conn_pair = (app_server_sender, app_server_receiver);
-        let app_permissions = AppPermissions {
-            reports: true,
-            routes: true,
-            send_funds: true,
-            config: true,
-        };
-        await!(connections_sender.send((app_permissions, app_server_conn_pair))).unwrap();
-
-
-        // The apps should receive the current node report as the first message:
-        let _to_app_message = await!(app_receiver0.next()).unwrap();
-        let _to_app_message = await!(app_receiver1.next()).unwrap();
-
-        // Send a request routes message through app0:
-        let request_routes = RequestRoutes {
-            request_id: Uid::from(&[3; UID_LEN]),
-            capacity: 250,
-            source: PublicKey::from(&[0xee; PUBLIC_KEY_LEN]),
-            destination: PublicKey::from(&[0xff; PUBLIC_KEY_LEN]),
-            opt_exclude: None,
-        };
-
-        await!(app_sender0.send(AppToAppServer::RequestRoutes(request_routes.clone()))).unwrap();
-
-        // RequestRoutes command should be forwarded to IndexClient:
-        let to_index_client_message = await!(index_client_receiver.next()).unwrap();
-        match to_index_client_message {
-            AppServerToIndexClient::RequestRoutes(received_request_routes) => 
-                assert_eq!(received_request_routes, request_routes),
-            _ => unreachable!(),
-        };
-
-        // IndexClient returns a response that is not related to any open request.
-        // This response will be discarded.
-        let client_response_routes = ClientResponseRoutes {
-            request_id: Uid::from(&[2; UID_LEN]),
-            result: ResponseRoutesResult::Failure,
-        };
-        await!(index_client_sender.send(IndexClientToAppServer::ResponseRoutes(client_response_routes))).unwrap();
-
-        // We shouldn't get an message at any of the apps:
-        assert!(app_receiver0.try_next().is_err());
-        assert!(app_receiver1.try_next().is_err());
-
-        // IndexClient returns a response corresponding to an open request:
-        let client_response_routes = ClientResponseRoutes {
-            request_id: Uid::from(&[3; UID_LEN]),
-            result: ResponseRoutesResult::Failure,
-        };
-        await!(index_client_sender.send(IndexClientToAppServer::ResponseRoutes(client_response_routes))).unwrap();
-
-        let to_app_message = await!(app_receiver0.next()).unwrap();
-        match to_app_message {
-            AppServerToApp::ResponseRoutes(response_routes) => {
-                assert_eq!(response_routes.request_id, Uid::from(&[3; UID_LEN]));
-                assert_eq!(response_routes.result, ResponseRoutesResult::Failure);
-            },
-            _ => unreachable!(),
-        }
-        // We shouldn't get an incoming message at app1:
-        assert!(app_receiver1.try_next().is_err());
-
-        // IndexClient again returns the same response.
-        // This time the response should be discarded, 
-        // because it does not correspond to any open request.
-        let client_response_routes = ClientResponseRoutes {
-            request_id: Uid::from(&[3; UID_LEN]),
-            result: ResponseRoutesResult::Failure,
-        };
-        await!(index_client_sender.send(IndexClientToAppServer::ResponseRoutes(client_response_routes))).unwrap();
-
-        // We shouldn't get an message at any of the apps:
-        assert!(app_receiver0.try_next().is_err());
-        assert!(app_receiver1.try_next().is_err());
-    }
-
-    #[test]
-    fn test_app_server_loop_index_request_routes() {
-        let mut thread_pool = ThreadPool::new().unwrap();
-        thread_pool.run(task_app_server_loop_request_routes(thread_pool.clone()));
-    }
-
-    // ------------------------------------------------
-    // ------------------------------------------------
-
-    async fn task_app_server_loop_request_send_funds<S>(mut spawner: S) 
-    where
-        S: Spawn + Clone + Send + 'static,
-    {
-
-        let (mut funder_sender, mut funder_receiver,
-             mut index_client_sender, mut index_client_receiver,
-             mut connections_sender, initial_node_report) = spawn_dummy_app_server(spawner.clone());
-
-        // Connect two apps:
-        let (mut app_sender0, app_server_receiver) = mpsc::channel(0);
-        let (app_server_sender, mut app_receiver0) = mpsc::channel(0);
-        let app_server_conn_pair = (app_server_sender, app_server_receiver);
-        let app_permissions = AppPermissions {
-            reports: true,
-            routes: true,
-            send_funds: true,
-            config: true,
-        };
-        await!(connections_sender.send((app_permissions, app_server_conn_pair))).unwrap();
-
-        let (mut app_sender1, app_server_receiver) = mpsc::channel(0);
-        let (app_server_sender, mut app_receiver1) = mpsc::channel(0);
-        let app_server_conn_pair = (app_server_sender, app_server_receiver);
-        let app_permissions = AppPermissions {
-            reports: true,
-            routes: true,
-            send_funds: true,
-            config: true,
-        };
-        await!(connections_sender.send((app_permissions, app_server_conn_pair))).unwrap();
-
-
-        // The apps should receive the current node report as the first message:
-        let _to_app_message = await!(app_receiver0.next()).unwrap();
-        let _to_app_message = await!(app_receiver1.next()).unwrap();
-
-        let pk_e = PublicKey::from(&[0xee; PUBLIC_KEY_LEN]);
-        let pk_f = PublicKey::from(&[0xff; PUBLIC_KEY_LEN]);
-
-        let user_request_send_funds = UserRequestSendFunds {
-            request_id: Uid::from(&[3; UID_LEN]),
-            route: FriendsRoute { public_keys: vec![pk_e.clone(), pk_f.clone()] },
-            invoice_id: InvoiceId::from(&[1; INVOICE_ID_LEN]),
-            dest_payment: 20,
-        };
-
-        await!(app_sender0.send(AppToAppServer::RequestSendFunds(user_request_send_funds.clone()))).unwrap();
-
-        // RequestRoutes command should be forwarded to IndexClient:
-        let to_funder_message = await!(funder_receiver.next()).unwrap();
-        match to_funder_message {
-            FunderIncomingControl::RequestSendFunds(received_user_request_send_funds) => 
-                assert_eq!(received_user_request_send_funds, user_request_send_funds),
-            _ => unreachable!(),
-        };
-
-        // Funder returns a response that is not related to any open request.
-        let response_received = ResponseReceived {
-            request_id: Uid::from(&[2; UID_LEN]),
-            result: ResponseSendFundsResult::Failure(pk_e.clone()),
-        };
-        await!(funder_sender.send(FunderOutgoingControl::ResponseReceived(response_received))).unwrap();
-
-        // We shouldn't get an message at any of the apps:
-        assert!(app_receiver0.try_next().is_err());
-        assert!(app_receiver1.try_next().is_err());
-
-        // Funder returns a response that corresponds to the open request:
-        let response_received = ResponseReceived {
-            request_id: Uid::from(&[3; UID_LEN]),
-            result: ResponseSendFundsResult::Failure(pk_e.clone()),
-        };
-        await!(funder_sender.send(FunderOutgoingControl::ResponseReceived(response_received.clone()))).unwrap();
-
-        let to_app_message = await!(app_receiver0.next()).unwrap();
-        match to_app_message {
-            AppServerToApp::ResponseReceived(obtained_response_received) => {
-                assert_eq!(obtained_response_received, response_received);
-            },
-            _ => unreachable!(),
-        }
-        // We shouldn't get an incoming message at app1:
-        assert!(app_receiver1.try_next().is_err());
-
-        // Funder again returns the same response,
-        // however, this time it will be discarded, because no open request
-        // has a matching id:
-        let response_received = ResponseReceived {
-            request_id: Uid::from(&[3; UID_LEN]),
-            result: ResponseSendFundsResult::Failure(pk_e),
-        };
-        await!(funder_sender.send(FunderOutgoingControl::ResponseReceived(response_received))).unwrap();
-
-        // We shouldn't get an message at any of the apps:
-        assert!(app_receiver0.try_next().is_err());
-        assert!(app_receiver1.try_next().is_err());
-    }
-
-    #[test]
-    fn test_app_server_loop_index_request_send_funds() {
-        let mut thread_pool = ThreadPool::new().unwrap();
-        thread_pool.run(task_app_server_loop_request_send_funds(thread_pool.clone()));
-    }
 }
