+++ components/app_server/src/tests/mod.rs
@@ -1,0 +1,7 @@
+mod utils;
+mod funder_command;
+mod index_client_command;
+mod all_apps_closed;
+mod request_routes;
+mod request_send_funds;
+mod two_apps;
